test/requirements.txt
stomp.py==8.*
pyyaml==6.*